Various minor fixes: spelling corrections; remove unused imports; correct type hierarchy; remove unnecessary casts; static access to static methods; remove unused private methods.

+++ joram/src/fr/dyade/aaa/jndi2/impl/StorageManager.java
@@ -64,8 +64,7 @@ public class StorageManager {
     if (contextCounterL == null) {
       contextCounter = 0;
     } else {
-      contextCounter = 
-        ((Long)contextCounterL).longValue();
+      contextCounter = contextCounterL.longValue();
     }
 
     // Load the context index
